requirements.txt
pandas
numpy
six
tqdm
biopython
tensorflow
xgboost
scikit-learn
joblib
Jinja2
matplotlib
seaborn
scipy
